+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Expect active development and potentially significant breaking changes in the `0
 - Removed AuthTokenHeaderMiddleware code and related tests from apollo-client [Issue #247](https://github.com/apollostack/apollo-client/issues/247)
 - Added named fragment support [Issue #80](https://github.com/apollostack/apollo-client/issues/80) and [PR #251](https://github.com/apollostack/apollo-client/pull/251).
 - Added basic guards to our Redux Store `subscribe` to prevent `broadcastQueries` from being called unnecessarily
+- Fixed polling leak issue that occured with multiple polling queries (https://github.com/apollostack/apollo-client/issues/248)
 
 ### v0.3.12